backend/requirements.txt
flask
flask-cors
flask-jwt-extended
werkzeug
psycopg2-binary
